backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
pydantic-settings==2.5.2
python-dotenv==1.0.1

# ML inference
torch==2.4.1
transformers==4.44.2
scikit-learn==1.5.2
joblib==1.4.2
numpy==1.26.4
pandas==2.2.3